chore: move defaults to a separate file
Currently the defalt values we use throughout the project are located in the same file
as the configuration logic. This commit resolves this so that the defaults live in a
separate file to be decoupled
